followed the tutorial to create mouselistener class https://www.youtube.com/watch?v=88oZT7Aum6s&list=PLtrSb4XxIVbp8AKuEAlwNXDxr99e3woGE&index=3&ab_channel=GamesWithGabe
i do not really understand why this is a singleton class.
Either make all variables static (this saves quite a bit of typing...)
or what i would prefere not make it singleton at all, so you could potential create to independent mouse Listener.
anyway i follow the tutorial, for the moment, and will refractor all this stuff anyway, as soon as i do something, i really want to use.

(only exception is the exception and the Stream which i used to avoid some ugly hardcoding